refactor softphone into hook and components
