+++ client/src/softphone/components/PopOutButton.jsx
@@ -1,0 +1,7 @@
+import { Button } from '@twilio-paste/core/button';
+
+export default function PopOutButton({ onClick }) {
+  return (
+    <Button variant="secondary" onClick={onClick}>Pop out</Button>
+  );
+}
